change reading vals to float

+++ Database/create_procs.sql
@@ -147,12 +147,12 @@ CREATE PROCEDURE addRead(
     dt DATETIME,    #timestamp of recording time
     p INT(4),       #project ID
     l INT(4),        #location ID
-    a DECIMAL(3,2), #sensor 1 value....
-    b DECIMAL(3,2),
-    c DECIMAL(3,2),
-    d DECIMAL(3,2),
-    e DECIMAL(3,2),
-    f DECIMAL(3,2)
+    a FLOAT, #sensor 1 value....
+    b FLOAT,
+    c FLOAT,
+    d FLOAT,
+    e FLOAT,
+    f FLOAT
     ) #...sensor 6 value
     COMMENT 'add a new reading to the Reading table, comes from LoRa Server'
     BEGIN
